+++ app/models/concerns/domain/force_delete.rb
@@ -19,6 +19,16 @@ module Concerns::Domain::ForceDelete # rubocop:disable Metrics/ModuleLength
     end
   end
 
+  def notification_template
+    if contact_emails_verification_failed.present?
+      'invalid_email'
+    elsif registrant.org?
+      'legal_person'
+    else
+      'private_person'
+    end
+  end
+
   def force_delete_scheduled?
     statuses.include?(DomainStatus::FORCE_DELETE)
   end
@@ -80,8 +90,8 @@ module Concerns::Domain::ForceDelete # rubocop:disable Metrics/ModuleLength
   end
 
   def cancel_force_delete
-    restore_statuses_before_force_delete
     remove_force_delete_statuses
+    restore_statuses_before_force_delete
     clear_force_delete_data
     self.force_delete_date = nil
     self.force_delete_start = nil
@@ -106,7 +116,7 @@ module Concerns::Domain::ForceDelete # rubocop:disable Metrics/ModuleLength
   end
 
   def soft_delete_dates(years)
-    self.force_delete_start = valid_to - years.years + 1.day
+    self.force_delete_start = valid_to - years.years
     self.force_delete_date = force_delete_start + Setting.expire_warning_period.days +
                              Setting.redemption_grace_period.days
   end
@@ -119,7 +129,7 @@ module Concerns::Domain::ForceDelete # rubocop:disable Metrics/ModuleLength
   end
 
   def preserve_current_statuses_for_force_delete
-    self.statuses_before_force_delete = statuses.clone
+    update(statuses_before_force_delete: statuses)
   end
 
   def restore_statuses_before_force_delete
@@ -128,9 +138,9 @@ module Concerns::Domain::ForceDelete # rubocop:disable Metrics/ModuleLength
   end
 
   def add_force_delete_statuses
-    statuses << DomainStatus::FORCE_DELETE
-    statuses << DomainStatus::SERVER_RENEW_PROHIBITED
-    statuses << DomainStatus::SERVER_TRANSFER_PROHIBITED
+    self.statuses |= [DomainStatus::FORCE_DELETE,
+                      DomainStatus::SERVER_RENEW_PROHIBITED,
+                      DomainStatus::SERVER_TRANSFER_PROHIBITED]
   end
 
   def remove_force_delete_statuses