+++ app/models/epp/domain.rb
@@ -612,6 +612,7 @@ class Epp::Domain < Domain
     statuses.delete(DomainStatus::SERVER_HOLD)
     statuses.delete(DomainStatus::EXPIRED)
     statuses.delete(DomainStatus::SERVER_UPDATE_PROHIBITED)
+    cancel_pending_delete
 
     save
   end