+++ test/mailers/previews/domain_delete_mailer_preview.rb
@@ -1,10 +1,11 @@
 class DomainDeleteMailerPreview < ActionMailer::Preview
   def self.define_forced_templates
-    DomainDeleteMailer.force_delete_templates.each do |template_name|
+    %w[private_person legal_person invalid_email].each do |template_name|
       define_method "forced_#{template_name}".to_sym do
-        DomainDeleteMailer.forced(domain: @domain,
-                                  registrar: @domain.registrar,
-                                  registrant: @domain.registrant,
+        domain = Domain.first
+        DomainDeleteMailer.forced(domain: domain,
+                                  registrar: domain.registrar,
+                                  registrant: domain.registrant,
                                   template_name: template_name)
       end
     end
@@ -12,26 +13,25 @@ class DomainDeleteMailerPreview < ActionMailer::Preview
 
   define_forced_templates
 
-  def initialize
-    @domain = Domain.first
-    super
-  end
-
   def confirmation_request
-    DomainDeleteMailer.confirmation_request(domain: @domain,
-                                            registrar: @domain.registrar,
-                                            registrant: @domain.registrant)
+    domain = Domain.first
+    DomainDeleteMailer.confirmation_request(domain: domain,
+                                            registrar: domain.registrar,
+                                            registrant: domain.registrant)
   end
 
   def accepted
-    DomainDeleteMailer.accepted(@domain)
+    domain = Domain.first
+    DomainDeleteMailer.accepted(domain)
   end
 
   def rejected
-    DomainDeleteMailer.rejected(@domain)
+    domain = Domain.first
+    DomainDeleteMailer.rejected(domain)
   end
 
   def expired
-    DomainDeleteMailer.expired(@domain)
+    domain = Domain.first
+    DomainDeleteMailer.expired(domain)
   end
 end