+++ test/system/admin_area/domains/force_delete_test.rb
@@ -42,6 +42,38 @@ class AdminAreaDomainForceDeleteTest < ApplicationSystemTestCase
       find(:css, '#soft_delete').set(true)
       click_link_or_button 'Force delete domain'
     end
+
+    @domain.reload
+    assert_equal @domain.notification_template, @domain.template_name
+  end
+
+  def test_uses_legal_template_if_registrant_org
+    @domain.registrant.update(ident_type: 'org')
+
+    assert_emails 0 do
+      visit edit_admin_domain_url(@domain)
+      find(:css, '#soft_delete').set(true)
+      click_link_or_button 'Force delete domain'
+    end
+
+    @domain.reload
+    assert_equal @domain.notification_template, @domain.template_name
+  end
+
+  def test_uses_legal_template_if_invalid_email
+    verification = @domain.contacts.first.email_verification
+    verification.update(verified_at: Time.zone.now - 1.day, success: false)
+
+    assert_equal @domain.notification_template, 'invalid_email'
+
+    assert_emails 0 do
+      visit edit_admin_domain_url(@domain)
+      find(:css, '#soft_delete').set(true)
+      click_link_or_button 'Force delete domain'
+    end
+
+    @domain.reload
+    assert_equal @domain.notification_template, @domain.template_name
   end
 
   def test_allows_to_skip_notifying_registrant_and_admin_contacts_by_email